+++ summoner-cli/src/Summoner/GhcVer.hs
@@ -5,6 +5,7 @@ and some useful functions for manipulation with them.
 module Summoner.GhcVer
        ( GhcVer (..)
        , Pvp (..)
+       , showGhcOutput
        , showGhcVer
        , parseGhcVer
        , latestLts
@@ -15,6 +16,8 @@ module Summoner.GhcVer
        ) where
 
 import Data.List (maximum, minimum)
+
+import Summoner.Text (intercalateMap)
 
 import qualified Text.Show as Show
 
@@ -27,6 +30,12 @@ data GhcVer
     | Ghc844
     | Ghc865
     deriving (Eq, Ord, Show, Enum, Bounded)
+
+showGhcOutput :: GhcVer -> Text
+showGhcOutput ghcVer = intercalateMap " " ($ ghcVer) [ showGhcVer
+                                                     , ("base-" <>) . baseVer
+                                                     , ("lts-" <>) . latestLts
+                                                     ]
 
 -- | Converts 'GhcVer' into dot-separated string.
 showGhcVer :: GhcVer -> Text